+++ FireObject.h
@@ -5,19 +5,19 @@ class FireObject
 public:
 	int myPlaneIndex = 1;
 
-	vector<float> myVectorFireObjectPosition = vector<float>(3);
-	vector<float> myVectorFireEmitterOffset = vector<float>(3);
+	vector<double> myVectorFireObjectPosition = vector<double>(3);
+	vector<double> myVectorFireEmitterOffset = vector<double>(3);
 
-	float myFireWaterRadius = 3.0f;
-	float myFireLiftNom = 77.0f;
-	float myFireLiftDenom = 5.7f;
-	float myFireDistDemon = 20.0f;
+	double myFireWaterRadius = 3.0;
+	double myFireLiftNom = 77.0;
+	double myFireLiftDenom = 5.7;
+	double myFireDistDemon = 20.0;
 
-	float myFireStrength = 1.0f;
+	double myFireStrength = 1.0;
 
 	double myLat = HSL_Data::coord_invalid;
 	double myLon = HSL_Data::coord_invalid;
-	float myElev = 0;
+	double myElev = 0;
 
 	XPLMDataRef myDrXPos = NULL;
 	XPLMDataRef myDrYPos = NULL;
@@ -39,13 +39,13 @@ public:
 
 
 
-	FireObject(HSL_PlugIn* pHSL, int inPlaneIndex, vector<float> inVectorFireObjectPosition, bool atCoordinates);
+	FireObject(HSL_PlugIn* pHSL, int inPlaneIndex, vector<double> inVectorFireObjectPosition, bool atCoordinates);
 
 	XPLMDataRef GetPlaneDataRef(std::string prefix, int planeIndex, std::string postfix);
 	void UpdateWorldCoordinates(void);
 	void SetPosition(void);
 	void Remove(void);
-	bool CheckWaterDrop(vector<float> inDropPosition, float inWaterAmount);
+	bool CheckWaterDrop(vector<double> inDropPosition, double inWaterAmount);
 	bool operator == (const FireObject& s);
 };
